requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
audioop-lts==0.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
decorator==4.4.2
discord.py==2.5.2
frozenlist==1.7.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
joblib==1.5.1
moviepy==1.0.3
multidict==6.6.3
numpy==2.2.4
pandas==2.2.3
pillow==10.4.0
proglog==0.1.11
propcache==0.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.3.0
yarl==1.20.1
